Fix DATADIR, shell wrapper, Makefile and configuration files
- Now 'make install' installs the engine in LIBEXECDIR; Piper libraries in LIBEXECDIR/libpiper and LIBEXECDIR/lib64; The shell wrapper is installed in $(BINDIR), becoming the main executable file of the application (the default installation puts it in /usr/local/bin directory, in the $PATH. If the user choose another path it's his/her responsibility to put it or not in the $PATH;

- make install-user -> creates $HOME/.config/$(APP_NAME) and installs data files and config files in that directory; This option is now required to make the application work correctly, even without installation;

- APP_NAME and INSTALL_PREFIX path are now injected by 'make install' in the shell wrapper, making the corresponding variables to point to the correct location;

- Defined APPLICATION_NAME in 'app.h' and makes the main function and the talker find the correct location of the config and data files; This was made to allow the c-engine executable work even without the shell wrapper, in case the user wants to open and execute the reader by command-line;

- Updated the logo in the shell wrapper.

+++ src/printer.c
@@ -1,5 +1,5 @@
 /*
-* MorceNOX CAT-Reader (C-Art Text Reader)™
+* MorceNOX Art-Reader™
 *
 * Copyright (C) 2026 Amilcar Antonio Mesquita Rizk <amilcar.rizk@gmail.com>
 *
@@ -375,7 +375,7 @@ void print_split_word(const char* str, int max_width) {
         size_t len = wcslen(wstr);
         size_t mid = len / 2;
 
-        if (mid * 4 + SCREEN_PADDING + 5 < max_width) {
+        if (mid * 4 + SCREEN_PADDING + 5 < (size_t)max_width) {
 
             // Instead of manual malloc/strncpy, use the helper above
             // to split the wide string and convert it back to UTF-8
